daemon/config: split portable bits out of config_linux.go
